otp at sign up

+++ client/src/Components/OTPComponent.jsx
@@ -1,18 +1,26 @@
 import { useEffect, useState } from 'react';
 import { Form } from 'react-bootstrap';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+import { Navigate } from 'react-router-dom';
 import { ToastContainer, toast } from 'react-toastify';
+import { createUser } from '../actions/userActions';
 import ButtonComponent from './ButtonComponent';
+import Signup from './Signup';
 
-const OTPComponent = ({ setEmailProp }) => {
+const OTPComponent = ({ setEmailProp, user, type, emailProp }) => {
   const [email, setEmail] = useState('');
   const [emailSent, setEmailSent] = useState(false);
   const [OTP, setOTP] = useState();
   const [buttonDisabled, setButtonDisabled] = useState(true);
   const dispatch = useDispatch();
+  const authenticated = useSelector((state) =>
+    state.user.authenticated
+      ? state.user.authenticated
+      : state.auth.authenticated
+  );
 
   useEffect(() => {
-    if (email) setButtonDisabled(false);
+    if (email && navigator.onLine) setButtonDisabled(false);
     else setButtonDisabled(true);
   }, [email]);
 
@@ -31,7 +39,7 @@ const OTPComponent = ({ setEmailProp }) => {
   const notifySuccess = (msg) => toast.success(msg, notificationProperties);
 
   const sendOTPMail = (email) => () => {
-    fetch(`http://localhost:7000/api/mails/sendOTP/${email}`)
+    fetch(`http://localhost:7000/api/mails/${type}/sendOTP/${email}`)
       .then((res) => {
         if (!res.ok) {
           return res.json().then((err) => {
@@ -44,11 +52,19 @@ const OTPComponent = ({ setEmailProp }) => {
         setEmailSent(true);
         notifySuccess(`OTP sent on ${email}`);
       })
-      .catch(() => {
-        notifyError('No user found with this email id.');
+      .catch((err) => {
+        notifyError(err.message);
         setEmailSent(false);
       });
   };
+
+  useEffect(() => {
+    if (type === 'signup') {
+      setEmail(emailProp);
+      dispatch(sendOTPMail(emailProp));
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   const sendOTP = (e) => {
     e.preventDefault();
@@ -60,12 +76,21 @@ const OTPComponent = ({ setEmailProp }) => {
     e.preventDefault();
     // eslint-disable-next-line eqeqeq
     if (localStorage.getItem('otp') == OTP) {
-      localStorage.clear();
-      setEmailProp(email);
+      if (type === 'forgotpassword') {
+        localStorage.clear();
+        setEmailProp(email);
+      } else {
+        notifySuccess('Email Verified');
+        dispatch(createUser(user));
+      }
     } else notifyError('Wrong OTP entered.');
   };
 
-  return (
+  if (authenticated) {
+    return <Navigate to={`/home`} />;
+  }
+
+  return type === 'forgotpassword' || emailSent ? (
     <>
       <Form onSubmit={emailSent ? confirmOTP : sendOTP}>
         <Form.Group className='mb-3'>
@@ -90,6 +115,8 @@ const OTPComponent = ({ setEmailProp }) => {
       </Form>
       <ToastContainer />
     </>
+  ) : (
+    <Signup />
   );
 };
